ref: I changed the code details to display not one but several animated objects on the full page so that the dimensions and other important control parameters for each of them are calculated according to the object's corresponding container, if necessary.
