+++ core/camera.js
@@ -1,12 +1,12 @@
 import * as THREE from 'three';
 
-const canvas = document.querySelector('#webgl');
-
-export const camera = new THREE.PerspectiveCamera(
-  75,
-  canvas.clientWidth / canvas.clientHeight,
-  0.1,
-  1000
-);
-
-camera.position.z = 5;
+export function createCamera(aspect) {
+  const camera = new THREE.PerspectiveCamera(
+    75,
+    aspect,
+    0.1,
+    1000
+  );
+  camera.position.z = 5;
+  return camera;
+}
