+++ core/controls.js
@@ -1,6 +1,7 @@
 import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls';
-import { camera } from './camera.js';
-import { renderer } from './renderer.js';
 
-export const controls = new OrbitControls(camera, renderer.domElement);
-controls.enableDamping = true;
+export function createControls(camera, renderer) {
+    const controls = new OrbitControls(camera, renderer.domElement);
+    controls.enableDamping = true;
+    return controls;
+}
