+++ core/renderer.js
@@ -1,15 +1,15 @@
 import * as THREE from 'three';
 
-const canvas = document.querySelector('#webgl');
+export function createRenderer(canvas) {
+  const renderer = new THREE.WebGLRenderer({
+    canvas,
+    antialias: true,
+    alpha: true
+  });
 
-export const renderer = new THREE.WebGLRenderer({
-  canvas: canvas,
-  antialias: true,
-  alpha: true
-});
+  const { width, height } = canvas.getBoundingClientRect();
+  renderer.setSize(width, height, false);
+  renderer.setPixelRatio(Math.min(window.devicePixelRatio, 2));
 
-const width = canvas.clientWidth;
-const height = canvas.clientHeight;
-
-renderer.setSize(width, height);
-renderer.setPixelRatio(Math.min(window.devicePixelRatio, 2));
+  return renderer;
+}
