+++ core/scene.js
@@ -1,6 +1,10 @@
 import * as THREE from 'three';
 
-export const scene = new THREE.Scene();
+export function createScene() {
+    const scene = new THREE.Scene();
 
-const ambientLight = new THREE.AmbientLight(0xffffff, 1.5);
-scene.add(ambientLight);
+    const ambientLight = new THREE.AmbientLight(0xffffff, 1.5);
+    scene.add(ambientLight);
+
+    return scene;
+}
