+++ index.html
@@ -8,7 +8,7 @@
   <body class="w-full h-full m-0 p-0 overflow-hidden bg-gray-900 text-white font-sans">
     <div class="flex w-full h-screen">
       <div class="flex justify-center items-center w-[70%] p-11 h-full">
-        <canvas id="webgl" class="!w-full !h-full border block"></canvas>
+        <canvas id="cube" class="!w-full !h-full border block"></canvas>
       </div>
 
       <aside class="w-[30%] h-full p-8 bg-[burlywood] overflow-y-auto flex flex-col gap-4">
